// projeto-angular/src/app/model/stats.ts
export class Stats {
    name: string = '';
    value: number = 0;
    total_kills: number = 0;
    total_deaths: number = 0;
    total_time_played: number = 0;
    total_wins: number = 0;
    total_kills_deagle: number = 0;
    total_kills_awp: number = 0;
    total_kills_ak47: number = 0;
    total_kills_m4a1: number = 0;
    total_kills_headshot: number = 0;
    total_kills_enemy_blinded: number = 0;
    total_rounds_map_de_vertigo: number = 0;
    total_rounds_map_de_inferno: number = 0;
    total_rounds_map_de_dust2: number = 0;
    total_rounds_map_de_train: number = 0;
    total_rounds_map_de_nuke: number = 0;

}

/*     total_kills: number = 0;
    total_deaths: number = 0;
    total_time_played: number = 0;
    total_wins: number = 0;
    total_kills_deagle: number = 0;
    total_kills_awp: number = 0;
    total_kills_ak47: number = 0;
    total_kills_m4a1: number = 0;
    total_kills_headshot: number = 0;
    total_kills_enemy_blinded: number = 0;
    total_rounds_map_de_vertigo: number = 0;
    total_rounds_map_de_inferno: number = 0;
    total_rounds_map_de_dust2: number = 0;
    total_rounds_map_de_train: number = 0;
    total_rounds_map_de_nuke: number = 0;

    */